requirements.txt
dj-database-url==0.4.1
Django==1.9.4
django-annoying==0.9.0
django-filter==0.13.0
djangorestframework==3.3.3
djoser==0.4.3
dwollaswagger>=1.0.0
gunicorn==19.4.5
httpie==0.9.3
Markdown==2.6.6
pbr==1.8.1
Pygments==2.1.3
requests==2.9.1
six==1.10.0
stevedore==1.9.0
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
whitenoise==3.0
